refactor(literature): move Mode 1 build orchestration + mesh extraction out of the route
Thin the search.ts route handler per review feedback on #824: the handler now
does request/response, validation and dispatch, and the business logic lives in
the service layer.

- literatureSearch.build.ts: buildSearchStrategy() — the Mode 1 multi-database
  build/count/seed-check/fixes loop, extracted from the route. It takes the cost
  accumulator and mutates it in place (not reassign) so an unforeseen throw still
  leaves the spend attributable, exactly as the hoisted `usage` did inline.
- literatureSearch.experts.ts: meshFromConcepts() + NOT_A_TOPIC — the MeSH->panel
  bridge, used by three branches. The findWcmExperts call and its per-branch error
  policy stay at the call sites (they differ: a re-count throws, a build swallows).

Behavior-neutral: the extracted loop is byte-identical to the inline one apart
from the in-place usage accumulation. tsc --noEmit clean; check:literature:pure
green. Route not driven live (worktree has no creds) — drive before merge.
