+++ controllers/literatureSearch.controller.ts
@@ -44,11 +44,14 @@
 // must never be one: the cap is what bounds the context window and therefore the bill.
 
 
-// This file is now the BARREL. The implementation lives in four focused modules, split by concern
-// so no single file is 2,000+ lines: strategy (query assembly + types), counting (arithmetic over
-// PubMed counts), records (fetch/shape/evidence tiers), llm (every Bedrock call). Consumers — the
-// API route and the check harnesses — import from here, so the split changed no call site.
+// This file is now the BARREL. The implementation lives in focused modules, split by concern so no
+// single file is 2,000+ lines: strategy (query assembly + types), counting (arithmetic over PubMed
+// counts), records (fetch/shape/evidence tiers), llm (every Bedrock call), experts (the MeSH→panel
+// bridge), build (the Mode 1 multi-database orchestration). Consumers — the API route and the check
+// harnesses — import from here, so the split changed no call site.
 export * from './literatureSearch.strategy'
 export * from './literatureSearch.counting'
 export * from './literatureSearch.records'
 export * from './literatureSearch.llm'
+export * from './literatureSearch.experts'
+export * from './literatureSearch.build'
